requirements.txt
Django==1.5.4
Markdown==2.3.1
MySQL-python==1.2.4
South==0.8.4
amqp==1.3.3
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.13
celery==3.1.7
distribute==0.7.3
django-celery==3.1.1
django-filter==0.7
django-kombu==0.9.4
django-redis==3.4.0
djangorestframework==2.3.10
kombu==3.0.8
pytz==2013.8
redis==2.8.0
wsgiref==0.1.2
